+++ lib/rugged/redis.rb
@@ -1,8 +1,3 @@
+require "rugged"
 require "rugged/redis/version"
-
-module Rugged
-  module Redis
-  end
-end
-
 require "rugged/redis/rugged_redis"